+++ Dracula/docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Improved the Death Coil card's tooltips.
+
 ## 1.1.5
 Released 19 May 2024.